refactor(storage): split god-files into domain files; dedup issue-ID; fix LoD
Break up the oversized storage files and the admin.go catch-all into focused,
domain-scoped files (all now well under the 500-line limit), with no behavior
change — existing storage tests cover every path:

- admin.go (750) -> issue_status.go / releases.go / alerts.go / settings.go /
  sourcemaps.go. Drops the dead scanAlert (only scanAlertWithProject is used).
- issues.go (760) -> issues.go (read) / issues_upsert.go (write) /
  issues_mutate.go / issues_counts.go.
- analytics.go (533) -> analytics.go (write/rollup) / analytics_query.go.

Dedup: add displayIssueID() helper consolidating the repeated
prefix/number-vs-legacy issue-ID branch across scanIssue, upsertIssue (x2) and
the event scan path.

Law of Demeter: config.go no longer reaches through s.oidc.Config().Issuer;
move buildEndSessionURL into the auth package and expose OIDCClient.LogoutURL()
and Issuer() so the handler is told what it needs.

Remove the three split files from the file-length allowlist so they are now
held to the 500-line limit.
